refactor(test): split oversized regression modules
